+++ drep/__init__.py
@@ -5,26 +5,74 @@ import os
 from Bio import SeqIO
 import shutil
 import multiprocessing
+import datetime
+import tqdm
 
-def run_cmd(cmd,dry=False,shell=True,quiet= True):
-    devnull = open(os.devnull, 'w')
-    if shell:
-        if not dry: 
-            if quiet:
-                call(cmd,shell=True,stdout=devnull, stderr=devnull)
-            else:
-                call(cmd,shell=True)
-        else: 
+# def run_cmd(cmd,dry=False,shell=True,quiet= True):
+#     devnull = open(os.devnull, 'w')
+#     if shell:
+#         if not dry:
+#             if quiet:
+#                 call(cmd,shell=True,stdout=devnull, stderr=devnull)
+#             else:
+#                 call(cmd,shell=True)
+#         else:
+#             print(cmd)
+#     else:
+#         if not dry:
+#             if quiet:
+#                 call(cmd,stdout=devnull, stderr=devnull)
+#             else:
+#                 call(cmd)
+#         else: print(' '.join(cmd))
+#     return
+
+def run_cmd(cmd, dry=False, shell=True, logdir=False):
+    # if dry, just print cmd
+    if dry:
+        if shell:
             print(cmd)
-    else: 
-        if not dry: 
-            if quiet:
-                call(cmd,stdout=devnull, stderr=devnull)
-            else:
-                call(cmd)
-        else: print(' '.join(cmd))
+        else:
+            print(' '.join(cmd))
+        return
+
+    # figure out what files you're going to store the output
+    if logdir == False:
+        devnull = open(os.devnull, 'w')
+        sto = devnull
+        ste = devnull
+    else:
+        uniq_filename = str(datetime.datetime.now().date()) + '_' + \
+            str(datetime.datetime.now().time()).replace(':', '.')
+        sto = open(os.path.join(logdir + uniq_filename + '.STDOUT'), 'w')
+        ste = open(os.path.join(logdir + uniq_filename + '.STDERR'), 'w')
+
+        # log the command
+        now = open(os.path.join(logdir + uniq_filename + '.CMD'), 'w')
+        if shell:
+            now.write(str(cmd) + '\n')
+        else:
+            now.write(' '.join(cmd) + '\n')
+        now.close()
+
+    # run the command
+    if shell:
+        call(cmd,shell=True,stdout=sto, stderr=ste)
+    else:
+        call(cmd,stdout=sto, stderr=ste)
     return
-    
+
+def thread_cmd_wrapper(tup):
+    run_cmd(*tup)
+
+def thread_cmds(cmds, dry=False, shell=False, logdir=False, t=10):
+    pool = multiprocessing.Pool(processes=t)
+    tups = [(cmd, dry, shell, logdir) for cmd in cmds]
+    pool.map(thread_cmd_wrapper, tups)
+    pool.close()
+    pool.join()
+    return
+
 def make_dir(outdirname,dry=False,overwrite=False):
     if dry:
         return
@@ -38,7 +86,7 @@ def make_dir(outdirname,dry=False,overwrite=False):
                             format(outdirname)
     else:
         os.makedirs(outdirname)
-        
+
 def clobber_dir(outdirname,dry=False,overwrite=False):
     if dry:
         return
@@ -51,14 +99,7 @@ def clobber_dir(outdirname,dry=False,overwrite=False):
                             format(outdirname)
     else:
         os.makedirs(outdirname)
-    
-def thread_cmds(cmds,t=10):
-    pool = multiprocessing.Pool(processes=t)
-    pool.map(run_cmd,cmds)
-    pool.close()
-    pool.join()
-    return
-    
+
 def fasta_length(fasta):
     total = 0
     for seq_record in SeqIO.parse(fasta, "fasta"):